Changed namespace & removed SetConfig from interface.
