+++ OptimizelySDK.NetStandard16/OptimizelySDK.NetStandard16.csproj
@@ -69,11 +69,11 @@
         <Compile Include="..\OptimizelySDK\Bucketing\UserProfileService.cs" />
         <Compile Include="..\OptimizelySDK\Bucketing\UserProfileUtil.cs" />
 		<Compile Include="..\OptimizelySDK\ProjectConfig.cs" />
-		<Compile Include="..\OptimizelySDK\DatafileManagement\DatafileProjectConfig.cs" />
-		<Compile Include="..\OptimizelySDK\DatafileManagement\ProjectConfigManager.cs" />
-		<Compile Include="..\OptimizelySDK\DatafileManagement\PollingProjectConfigManager.cs" />
-		<Compile Include="..\OptimizelySDK\DatafileManagement\HttpProjectConfigManager.cs" />
-		<Compile Include="..\OptimizelySDK\DatafileManagement\FallbackProjectConfigManager.cs">
+		<Compile Include="..\OptimizelySDK\Config\DatafileProjectConfig.cs" />
+		<Compile Include="..\OptimizelySDK\Config\ProjectConfigManager.cs" />
+		<Compile Include="..\OptimizelySDK\Config\PollingProjectConfigManager.cs" />
+		<Compile Include="..\OptimizelySDK\Config\HttpProjectConfigManager.cs" />
+		<Compile Include="..\OptimizelySDK\Config\FallbackProjectConfigManager.cs">
 		  <Link>AtomicProjectConfigManager.cs</Link>
 		</Compile>
     </ItemGroup>
