+++ OptimizelySDK.Tests/AudienceConditionsTests/ConditionsTest.cs
@@ -17,7 +17,7 @@
 using Moq;
 using NUnit.Framework;
 using OptimizelySDK.AudienceConditions;
-using OptimizelySDK.DatafileManagement;
+using OptimizelySDK.Config;
 using OptimizelySDK.Entity;
 using OptimizelySDK.Logger;
