+++ OptimizelySDK.Tests/DatafileManagement Tests/HttpProjectConfigManagerTest.cs
@@ -16,7 +16,7 @@
 
 using Moq;
 using NUnit.Framework;
-using OptimizelySDK.DatafileManagement;
+using OptimizelySDK.Config;
 using OptimizelySDK.Logger;
 using System;
 using System.Diagnostics;
