+++ OptimizelySDK.Tests/DatafileManagement Tests/PollingProjectConfigManagerTest.cs
@@ -16,7 +16,7 @@
 
 using Moq;
 using NUnit.Framework;
-using OptimizelySDK.DatafileManagement;
+using OptimizelySDK.Config;
 using OptimizelySDK.Logger;
 using OptimizelySDK.Tests.DatafileManagementTests;
 using System;
