+++ OptimizelySDK.Tests/EventTests/EventBuilderTest.cs
@@ -24,7 +24,7 @@ using System;
 using NUnit.Framework;
 using OptimizelySDK.Bucketing;
 using OptimizelySDK.Utils;
-using OptimizelySDK.DatafileManagement;
+using OptimizelySDK.Config;
 
 namespace OptimizelySDK.Tests.EventTests
 {
