+++ OptimizelySDK.Tests/NotificationTests/NotificationCenterTests.cs
@@ -16,7 +16,7 @@
 
 using Moq;
 using NUnit.Framework;
-using OptimizelySDK.DatafileManagement;
+using OptimizelySDK.Config;
 using OptimizelySDK.Entity;
 using OptimizelySDK.ErrorHandler;
 using OptimizelySDK.Event;
