+++ OptimizelySDK/Config/HttpProjectConfigManager.cs
@@ -21,7 +21,7 @@ using System.Collections.Generic;
 using System.Linq;
 using System.Threading.Tasks;
 
-namespace OptimizelySDK.DatafileManagement
+namespace OptimizelySDK.Config
 {
     public class HttpProjectConfigManager : PollingProjectConfigManager
     {
